io_grid, esri ascii format: bug fix, refers to http://sourceforge.net/p/saga-gis/bugs/163/

+++ saga-gis/src/modules_io/grid/io_grid/esri_arcinfo.cpp
@@ -597,7 +597,7 @@ bool CESRI_ArcInfo_Export::On_Execute(void)
 					fprintf(Stream.Get_Stream(), " ");
 				}
 
-				fprintf(Stream.Get_Stream(), Write_Value(pGrid->asFloat(x, y), Precision, bComma).b_str());
+				fprintf(Stream.Get_Stream(), Write_Value(pGrid->asDouble(x, y), Precision, bComma).b_str());
 			}
 
 			fprintf(Stream.Get_Stream(), "\n");
